pyproject.toml
[project]
name = "fitness"
version = "0.1.0"
description = ""
authors = [
    {name = "daniil_linevych21",email = "dan4224rit@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "django-types (>=0.22.0,<0.23.0)",
    "django-stubs-ext (>=5.2.7,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg (>=3.2.12,<4.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "django-filter (>=25.2,<26.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "faker (>=38.2.0,<39.0.0)",
    "openai (>=2.8.1,<3.0.0)",
    "anthropic (>=0.75.0,<0.76.0)",
    "celery (>=5.6.0,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)"
]

[tool.poetry]
packages = [{include = "fitness", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"
pytest-django = "^4.11.1"
autopep8 = "^2.3.2"
flake8 = "^7.3.0"
pre-commit = "^4.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
